Feature: Created Separate classes and included Interfaces for CityModel and CountryModel
