+++ City/CitiesContext.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.ComponentModel.DataAnnotations;
+using System.Data.Entity;
+
+namespace Cities
+{
+    public class CitiesContext : DbContext
+    {
+        public DbSet<CountryEntity> Countries { get; set; }
+        public DbSet<CityEntity> Cities { get; set; }
+    }
+}
