+++ City/CityEntity.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Data.Entity;
+using System.ComponentModel.DataAnnotations;
+
+namespace Cities
+{
+    public class CityEntity : CityModel
+    {
+        [Key]
+        public int CityID { get; set; }
+
+        public virtual CountryEntity Country { get; set; }
+    }
+}
