+++ City/CityModel.cs
@@ -4,39 +4,19 @@ using System.Data.Entity;
 
 namespace Cities
 {
-    public class CityModelImport
 
+    public interface ICityModel
     {
-        public string City_name { get; set; }
-        public string City_ascii { get; set; }
-        public double Lat { get; set; }
-        public double Lng { get; set; }
-        public string Country { get; set; }
-        public string ISO2 { get; set; }
-        public string ISO3 { get; set; }
-        public string Admin_name { get; set; }
-        public string Capital { get; set; }
-        public double Population { get; set; }
-        public float Id { get; set; }
+        public CityModel AddCities();
     }
-
-    public class CountryModel
+       
+    public class CityModel : ICityModel
     {
-        public string Name { get; set; }
-        public string ISO2 { get; set; }
-        public string ISO3 { get; set; }
-    }
+        public CityModel AddCities() {
 
-    public class CountryEntity : CountryModel
-    {
-        [Key]
-        public int CountryID { get; set; }
-
-        public virtual List<CityEntity> Cities { get; set; }
-    }
-
-    public class CityModel
-    {
+            var db = DBConn.conn();
+            return null;
+        }
         public string City_name { get; set; }
         public string City_ascii { get; set; }
         public double Lat { get; set; }
@@ -47,17 +27,5 @@ namespace Cities
         public int CountryId { get; set; }
     }
 
-    public class CityEntity : CityModel
-    {
-        [Key]
-        public int CityID { get; set; }
-
-        public virtual CountryEntity Country { get; set; }
-    }
-
-    public class CitiesContext : DbContext
-    {
-        public DbSet<CountryEntity> Countries { get; set; }
-        public DbSet<CityEntity> Cities { get; set; }
-    }
+    
 }
