+++ City/CityModelImport.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Cities
+{
+    public class CityModelImport
+    {
+
+            public string City_name { get; set; }
+            public string City_ascii { get; set; }
+            public double Lat { get; set; }
+            public double Lng { get; set; }
+            public string Country { get; set; }
+            public string ISO2 { get; set; }
+            public string ISO3 { get; set; }
+            public string Admin_name { get; set; }
+            public string Capital { get; set; }
+            public double Population { get; set; }
+            public float Id { get; set; }
+        
+
+    }
+}
